pipe overhaul

+++ Commands.cpp
@@ -218,9 +218,15 @@ void PipeCommand::execute() {
         setpgrp();      // make sure that the child gets a different GROUP ID
 
         pid_t pid1, pid2;
-        bool success = Pipe(&pid1, &pid2);
+        bool success = PipeCommand::Pipe(&pid1, &pid2);
         if (!sucess) {
             // kill this process and it's children
+            pid_t gpid = getpgid(CURR_FORK_CHILD_RUNNING);
+            if (gpid < 0) {
+                perror("smash error: getgpid failed");
+                exit(0);
+            }
+
             if (killpg(gpid, SIGINT) < 0) {
                 perror("smash error: killpg failed");
                 exit(0);
